moved colormap into its own Styelr file,  you can now pass in an array for a color map
